03_Implementacao/SemanticSearch/test/requirements.txt
numpy
matplotlib
faiss-cpu
ollama
requests
flask
PyPDF2
python-docx
markdown